+++ R/RegionDefinitions.R
@@ -136,12 +136,24 @@ makeNullRegionDefinition <- function(regionLength) {
 
 #' IMGT unique numbering schemes
 #'
-#' Definitions of the CDR and FWR, according to the IMGT unique numbering scheme.
+#' Sequence region definitions according to the IMGT unique numbering scheme.
 #'
-#' @format A \link{RegionDefinition} object containing:
+#' @format A \link{RegionDefinition} object defining:
 #' \itemize{
-#'   \item  \code{IMGT_V}:                     Grouped CDR and FWR V-segment regions excluding CDR3.
-#'   \item  \code{IMGT_V_BY_REGIONS}:          Individual CDR and FWR V-segment regions excluding CDR3.
+#'   \item  \code{IMGT_V}:              The IMGT numbered V segment up to position nucleotide 312.
+#'                                      This definition combines the CDR1 and CDR2 into a single CDR region,
+#'                                      and FWR1, FWR2 and FWR3 into a single FWR region. CDR3 and FWR4 are
+#'                                      excluded as they are downstream of nucleotide 312.
+#'   \item  \code{IMGT_V_BY_CODONS}:    The IMGT numbered V segment up to position nucleotide 312.
+#'                                      This definition treats each codon, from codon 1 to codon 104, as a 
+#'                                      distinct region.
+#'   \item  \code{IMGT_V_BY_REGIONS}:   The IMGT numbered V segment up to position nucleotide 312.
+#'                                      This defines separate regions for each of CDR1, CDR2,
+#'                                      FWR1, FWR2 and FWR3. CDR3 and FWR4 are
+#'                                      excluded as they are downstream of nucleotide 312.
+#'   \item  \code{IMGT_V_BY_SEGMENTS}:  The IMGT numbered V segment up to position nucleotide 312.
+#'                                      This definition has no subdivisons and treats the entire V segment
+#'                                      as a single region.
 #' }
 #' 
 #' @references
@@ -158,7 +170,14 @@ NULL
 #' @rdname  IMGT_SCHEMES
 NULL
 
+#' @name    IMGT_V_BY_CODONS
+#' @rdname  IMGT_SCHEMES
+NULL
+
 #' @name    IMGT_V_BY_REGIONS
 #' @rdname  IMGT_SCHEMES
 NULL
 
+#' @name    IMGT_V_BY_SEGMENTS
+#' @rdname  IMGT_SCHEMES
+NULL