+++ R/Shazam.R
@@ -141,7 +141,8 @@
 #'                          approx convolve weighted.mean p.adjust
 #'                          dbeta pbeta qbeta rbeta optim optimize
 #'                          dnorm pnorm runif dgamma pgamma uniroot na.exclude
-#' @importFrom  stringi     stri_dup stri_flatten stri_join stri_length
+#' @importFrom  stringi     stri_dup stri_flatten stri_join stri_length 
+#'                          stri_sub stri_sub<-
 #'                          stri_count_boundaries stri_count_regex 
 #'                          stri_extract_all_regex stri_extract_first_regex  
 #'                          stri_replace_all_regex stri_replace_first_regex